Connect product CRUD to PostgreSQL

+++ src/main/java/com/niva/shopcartpro/dto/ProductResponse.java
@@ -1,12 +1,14 @@
 package com.niva.shopcartpro.dto;
 
+import java.math.BigDecimal;
+
 public class ProductResponse {
 
     private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
 
-    public ProductResponse(Long id, String name, double price) {
+    public ProductResponse(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -20,7 +22,7 @@ public class ProductResponse {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
